Add Column class and public col() function
