+++ pointblank/__init__.py
@@ -10,7 +10,8 @@ except PackageNotFoundError:  # pragma: no cover
 
 # Import objects from the module
 from pointblank.tf import TF
+from pointblank.column import col
 from pointblank.validate import Validate, load_dataset
 from pointblank.thresholds import Thresholds
 
-__all__ = ["TF", "Validate", "Thresholds", "load_dataset"]
+__all__ = ["TF", "Validate", "Thresholds", "col", "load_dataset"]
